temporarily removed admin notices #27
The check for when to display which admin notice after plugin
activation didn’t take other locales than German into account. But
since the rest of the override process is locale-agnostic as long as
there is a valid .mo file for a language, it seemed to make sense to
remove the on-boarding process temporarily, until it can be revisited
and refined.

closes #27 when merged